+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Controller/StudentController.java
@@ -33,7 +33,7 @@ public class StudentController {
     }
 
     @PostMapping("/deleteStudent/{userId}")
-    public ResponseEntity<String> deleteUser(@PathVariable("userId") int userId) {
+    public ResponseEntity<String> deleteUser(@PathVariable("userId") Long userId) {
         studentService.deleteStudent(userId);
         return ResponseEntity.ok("Student deleted");
     }
